Refactored code, bumped version to 1.0 final. NEW FEATURE: REQUESTS
/radio request <song>
Adds a request for a song.
If there is no DJ, it will not accept requests.

/radio reqlist <num, del <all, num>>
No parameters: returns the first 5 requests, if any.
del <num,all> [REQUIRES ADMIN OR CURRENT DJ]: Deletes the request of that number or all requests.
If there is no DJ, it will not list requests.